wip: add test for lookup generation
